pyproject.toml
[project]
name = "agentboiler"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.97.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.0",
    "pydantic-ai>=1.87.0",
    "pydantic-settings>=2.14.0",
    "PyJWT[crypto]>=2.10.0",
    "python-dotenv>=1.2.2",
    "stripe[async]==15.1.0",
    "supabase>=2.29.0",
    "uvicorn[standard]>=0.46.0",
]
